refactor(desktop): reorganize settings tabs — General / Ticker / Account
Restructure settings sub-tabs from Appearance/Behavior/Account to
General/Ticker/Account for a more standard, intuitive layout:

- General: theme, UI scale, window options (always on top, taskbar,
  in-app ticker), and startup — the stuff you set once and forget
- Ticker: all ticker customization (show, rows, density, speed,
  hover, gap, item order, chip colors) — the stuff you tweak
- Account: unchanged (sign in/out, plan, updates, reset)

Files:
- New GeneralSettings.tsx (merges old Appearance theme/scale + Behavior window/startup)
- New TickerSettings.tsx (ticker layout/playback/style from old Appearance)
- Deleted AppearanceSettings.tsx and BehaviorSettings.tsx
- Updated Sidebar SettingsTab type and SETTINGS_SUBS
- Updated SettingsPanel tab routing
- Migration guard for persisted 'appearance'/'behavior' tab values
- Bumped version to 0.5.2
