+++ desktop/src-tauri/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "scrollr-desktop"
-version = "0.5.1"
+version = "0.5.2"
 edition = "2021"
 
 [lib]
